test(worktree): split real-git integration tests into sibling .integration.test.js (refs #3652)
Move the two real-git worktree round-trip tests out of
worktree-manager.test.js into a new worktree-manager.integration.test.js
sibling so the git-spawning tests no longer run in the test:quick TDD loop.
The new file is matched by the existing tests/**/*.integration.test.js glob in
INTEGRATION_INCLUDE, so it runs under test:integration and npm test but is
excluded from test:quick. The mock-only test file retains all pure-unit tests.
